% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_cumulative_hazard.R
\name{Average.MC.cumhaz}
\alias{Average.MC.cumhaz}
\title{Average several Kaplan-Meier (or Nelson-Aalen/Breslow) curves}
\usage{
Average.MC.cumhaz(MClist)
}
\arguments{
\item{MClist}{a list of 'cumhaz.matrix' objects (see coun.arm.events), as for instance obtained via a simulation/resampling procedure, and used to produce a single averaged Kaplan-Meier (or Nelson-Aalen/Breslow) curve}
}
\description{
Average several Kaplan-Meier (or Nelson-Aalen/Breslow) curves
}
\examples{

help("st4sas")

}
