% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxoptim.R
\name{CollectLoglikResults}
\alias{CollectLoglikResults}
\title{collects results from an optimized log likelihood.}
\arguments{
\item{obj}{an output from the 'maxLik' function}
}
\description{
extracts basic optimization output and computes meaningful summaries like MLE estimates, its covariance, normally approximates 95\% confidence intervals, log likelihood deviance, and the AIC.
}
