% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_cumulative_hazard.R
\name{count.arm.events}
\alias{count.arm.events}
\title{Time-events (points) count of a (stratified) Kaplan-Meier (Nelson-Aalen/Breslow) estimate.}
\usage{
count.arm.events(cumhaz.matrix)
}
\arguments{
\item{cumhaz.matrix}{a data.frame with information on Kaplan-Meier (Nelson-Aalen/Breslow) values at each event-time point, along point-wise confidence interval (CI) values and treatment/control indicator. By convention such information must be arranged in columns with name 'km' (survival probability), 'time' (event-time), 'km_l' (lower CI value for 'km'), 'km_u' (upper CI value for 'km'), 'trt' (binary indicator for a treatment/control grouping, can be NA).}
}
\description{
Time-events (points) count of a (stratified) Kaplan-Meier (Nelson-Aalen/Breslow) estimate.
}
\details{
basic information can be, for instance, computed via 'survfit' from package 'survival'.
}
