% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxoptim.R
\name{coxlik}
\alias{coxlik}
\title{Cox partial likelihood maximization using sufficient statistics formulation.}
\arguments{
\item{y}{(vector valued) sufficient statistic. It automatically enters the Cox log likelihood nominator}

\item{x}{raw covariates data (one observation per individual) in matrix form}

\item{str}{categorical vector indicating several strata to be used in a stratified Cox refression.}
}
\description{
Performs usual Cox partial likelihood optimization (using package 'maxLik') but uses sufficient statistics where possible (e.g. in the log likelihood numerator) instead of raw data, that is instead used in the covariates-based risk-set (log likelihood denominator). This setting can be useful in simulaiton settings where the numerator can be kept fixed and denominator raw data is simulated instead. For instance one might recover summary information on number-of-events in a treatment arm, or number-of-events times blood pressure values, but not raw data on treatment assignement and blood pressure values. Under such incomplete data situation one might still run a Cox regression noting that the summary information is a sufficient statistic in the Cox likelihood and trying to simulate raw covariate information in the denominator, akin to [1] or [2].
}
\references{
[1] Borgan, Ø. and Keogh, R. (2015). Nested case–control studies: should one break the matching? Lifetime Data Analysis, 21(4):517–541.

[2] Keogh, R. H., Seaman, S. R., Bartlett, J. W., and Wood, A. M. (2018). Multiple imputation of missing data in nested case-control and case-cohort studies. Biometrics.
}
